Crop_Recommendation/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
Pillow==10.0.0
transformers 
torch 
scikit-learn 
pandas 
numpy 
matplotlib 
requests 
plotly
joblib
